requirements.txt
injector==0.21.0
openai==1.97.1
openai-agents==0.2.4

python-dotenv==1.0.1

Flask==3.0.2
Flask-WTF==1.2.2
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
flask-cors==5.0.1
flask-login==0.6.3

unstructured==0.17.2
openpyxl==3.1.5
python-docx==1.1.2
python-pptx==1.0.2

spacy==3.8.7
transformers
sentence-transformers==4.0.2
faiss-cpu==1.10.0
rank_bm25==0.2.2

langchain==0.2.1
langchain-community==0.2.1
langchain-experimental==0.0.60

pytest==8.3.4

psycopg2==2.9.10
langchain-openai

langgraph==0.3.21

lark==1.2.2
umap-learn==0.5.7
marshmallow==3.26.1
redis==6.4.0
langfuse==2.60.9
openinference-instrumentation-openai
opentelemetry-sdk
opentelemetry-exporter-otlp
openinference-instrumentation-langchain==0.1.49

pydantic-evals==0.6.2
boto3==1.40.21
arize==7.50.0
celery==5.5.3
jieba==0.42.1
uv==0.9.4

pyJWT
